boot: final round of fixes
* fix ALIGN macro
* match memory address calculation to kexec-tools
 * 0x8000 text segment
 * validate zImage, ensure file is not truncated before attempting to
 boot
 * add decompression space at the end of the zImage
* handle errors better
* set initrd start/end addresses in dtb

+++ boot/util.h
@@ -16,6 +16,6 @@
  */
 
 struct device_config *get_cur_device(struct global_config *cfg);
-void *load_dtb(struct global_config *cfg, struct device_config *dev, int *dtbsz);
-void *load_file(struct global_config *cfg, char *file, int *sz);
-void *load_overlay(struct global_config *cfg, struct overlay_cfg *overlay, int *sz);
+void *load_dtb(struct global_config *cfg, struct device_config *dev, off_t *dtbsz);
+void *load_file(struct global_config *cfg, char *file, off_t *sz);
+void *load_overlay(struct global_config *cfg, struct overlay_cfg *overlay, off_t *sz);
